+++ papao-api/src/main/java/com/papaolabs/api/domain/model/Stats/Stats.java
@@ -1,0 +1,4 @@
+package com.papaolabs.api.domain.model.Stats;
+
+public class Stats {
+}
