+++ papao-api/src/main/java/com/papaolabs/api/domain/service/AnimalServiceImpl.java
@@ -1,11 +1,11 @@
-package com.papaolabs.api.service;
+package com.papaolabs.api.domain.service;
 
-import com.papaolabs.api.controller.dto.AnimalRequest;
-import com.papaolabs.api.restapi.AnimalApiClient;
-import com.papaolabs.api.restapi.dto.AnimalApiResponse;
-import com.papaolabs.api.restapi.dto.AnimalKindApiResponse;
-import com.papaolabs.api.restapi.dto.RegionApiResponse;
-import com.papaolabs.api.restapi.dto.ShelterApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.AnimalApiClient;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.AnimalApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.AnimalKindApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.RegionApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.ShelterApiResponse;
+import com.papaolabs.api.interfaces.v1.dto.AnimalRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
