+++ papao-api/src/main/java/com/papaolabs/api/infrastructure/persistence/jpa/JpaConfiguration.java
@@ -1,0 +1,4 @@
+package com.papaolabs.api.infrastructure.persistence.jpa;
+
+public class JpaConfiguration {
+}
