+++ papao-api/src/main/java/com/papaolabs/api/infrastructure/persistence/restapi/AnimalApiClientFallbackFactory.java
@@ -1,10 +1,10 @@
-package com.papaolabs.api.restapi;
+package com.papaolabs.api.infrastructure.persistence.restapi;
 
 import com.netflix.hystrix.exception.HystrixRuntimeException;
-import com.papaolabs.api.restapi.dto.AnimalApiResponse;
-import com.papaolabs.api.restapi.dto.AnimalKindApiResponse;
-import com.papaolabs.api.restapi.dto.RegionApiResponse;
-import com.papaolabs.api.restapi.dto.ShelterApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.AnimalApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.AnimalKindApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.RegionApiResponse;
+import com.papaolabs.api.infrastructure.persistence.restapi.dto.ShelterApiResponse;
 import feign.Param;
 import feign.hystrix.FallbackFactory;
 import org.springframework.util.MultiValueMap;
