+++ papao-api/src/main/java/com/papaolabs/api/infrastructure/persistence/restapi/dto/AnimalKindApiResponse.java
@@ -1,4 +1,4 @@
-package com.papaolabs.api.restapi.dto;
+package com.papaolabs.api.infrastructure.persistence.restapi.dto;
 
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
