+++ papao-api/src/main/java/com/papaolabs/api/infrastructure/persistence/restapi/dto/ShelterApiResponse.java
@@ -1,4 +1,4 @@
-package com.papaolabs.api.restapi.dto;
+package com.papaolabs.api.infrastructure.persistence.restapi.dto;
 
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -9,7 +9,7 @@ import java.util.List;
 @Data
 @XmlRootElement(name = "response")
 @RequiredArgsConstructor
-public class RegionApiResponse {
+public class ShelterApiResponse {
     private Header header;
     private Body body;
 
@@ -25,13 +25,12 @@ public class RegionApiResponse {
 
         @Data
         public static class Items {
-            List<RegionItemDTO> item;
+            List<ShelterItemDTO> item;
 
             @Data
-            public static class RegionItemDTO {
-                private String uprCd;
-                private String orgCd;
-                private String orgdownNm;
+            public static class ShelterItemDTO {
+                private String careRegNo;
+                private String careNm;
             }
         }
     }
