+++ papao-api/src/main/java/com/papaolabs/api/interfaces/v1/controller/V1DashboardController.java
@@ -1,8 +1,8 @@
-package com.papaolabs.api.controller;
+package com.papaolabs.api.interfaces.v1.controller;
 
-import com.papaolabs.api.controller.dto.AnimalRequest;
-import com.papaolabs.api.controller.dto.StatsDTO;
-import com.papaolabs.api.service.StatsService;
+import com.papaolabs.api.interfaces.v1.dto.AnimalRequest;
+import com.papaolabs.api.interfaces.v1.dto.StatsDTO;
+import com.papaolabs.api.domain.service.StatsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
