+++ papao-api/src/main/java/com/papaolabs/api/interfaces/v1/presenter/V1ShowPagePresenter.java
@@ -1,0 +1,4 @@
+package com.papaolabs.api.interfaces.v1.presenter;
+
+public class V1ShowPagePresenter {
+}
