--- papao-api/src/main/java/com/papaolabs/api/service/StatsService.java
@@ -1,8 +1,0 @@
-package com.papaolabs.api.service;
-
-import com.papaolabs.api.controller.dto.StatsDTO;
-
-public interface StatsService {
-
-    StatsDTO getStats(String beginDate, String endDate);
-}
